refactor(label): rename to rh-tag
