--- elements/rh-label/demo/rh-label.html
@@ -1,31 +1,0 @@
-<link rel="stylesheet" href="./demo.css"/>
-<script type="module" src="./rh-label.js"></script>
-
-
-  <h2 slot="header">Filled Color</h2>
-  <rh-label color="blue">Blue</rh-label>
-  <rh-label color="green">Green</rh-label>
-  <rh-label color="orange">Orange</rh-label>
-  <rh-label color="red"><span>Red <span class="visually-hidden-class">Hat</span></span></rh-label>
-  <rh-label color="purple">Purple</rh-label>
-  <rh-label color="cyan">Cyan</rh-label>
-  <rh-label>Grey</rh-label>
-  <div class="example">
-      Example:
-      <pfe-codeblock><div codeblock-container code-language="html"><code>&lt;rh-label color="blue"&gt;Blue&lt;/rh-label&gt;</code></div></pfe-codeblock>
-  </div>
-
-  <h2 slot="header">With Icon</h2>
-  <div>
-    <rh-label icon="circle-info" icon-set="fas">Default</rh-label>
-    <rh-label color="blue" icon="circle-info" icon-set="fas">Blue label</rh-label>
-    <rh-label color="green">Green label <svg slot="icon" height="1em" width: "1em" fill="currentColor" viewBox="0 0 512 512" aria-hidden="true" role="img"><path d="M256 8C119.043 8 8 119.083 8 256c0 136.997 111.043 248 248 248s248-111.003 248-248C504 119.083 392.957 8 256 8zm0 110c23.196 0 42 18.804 42 42s-18.804 42-42 42-42-18.804-42-42 18.804-42 42-42zm56 254c0 6.627-5.373 12-12 12h-88c-6.627 0-12-5.373-12-12v-24c0-6.627 5.373-12 12-12h12v-64h-12c-6.627 0-12-5.373-12-12v-24c0-6.627 5.373-12 12-12h64c6.627 0 12 5.373 12 12v100h12c6.627 0 12 5.373 12 12v24z"></path></svg></rh-label>
-    <rh-label color="orange" icon="circle-info" icon-set="fas">Orange label</rh-label>
-    <rh-label color="red" icon="circle-info" icon-set="fas">Red label</rh-label>
-    <rh-label color="purple" icon="circle-info" icon-set="fas">Purple label <svg slot="icon" height="1em" width: "1em" fill="currentColor" viewBox="0 0 512 512" aria-hidden="true" role="img"><path d="M256 8C119.043 8 8 119.083 8 256c0 136.997 111.043 248 248 248s248-111.003 248-248C504 119.083 392.957 8 256 8zm0 110c23.196 0 42 18.804 42 42s-18.804 42-42 42-42-18.804-42-42 18.804-42 42-42zm56 254c0 6.627-5.373 12-12 12h-88c-6.627 0-12-5.373-12-12v-24c0-6.627 5.373-12 12-12h12v-64h-12c-6.627 0-12-5.373-12-12v-24c0-6.627 5.373-12 12-12h64c6.627 0 12 5.373 12 12v100h12c6.627 0 12 5.373 12 12v24z"></path></svg></rh-label>
-    <rh-label color="cyan" icon="circle-info" icon-set="fas">Cyan label</rh-label>
-  </div>
-  <div class="example">
-    Example:
-    <pfe-codeblock><div codeblock-container code-language="html"><code>&lt;rh-label color="blue" icon="circle-info" icon-set="fas"&gt;Blue&lt;/rh-label&gt;</code></div></pfe-codeblock>
-</div>
