--- elements/rh-label/demo/rh-label.js
@@ -1,5 +1,0 @@
-import '@rhds/elements/rh-label/rh-label.js';
-
-const root = document.querySelector('[data-demo="rh-label"]')?.shadowRoot ?? document;
-
-root.querySelector('rh-label');
