--- elements/rh-label/docs/rh-label.md
@@ -1,40 +1,0 @@
-{% renderOverview %}
-  <rh-label color="blue">Blue</rh-label>
-{% endrenderOverview %}
-
-{% band header="Usage" %}
-```html
-<rh-label color="blue">Blue</rh-label>
-```
-{% endband %}
-
-{% band header="Variants" %}
-### With Icon
-<rh-label color="blue" icon="web-icon-alert-danger">Blue label</rh-label>
-
-```html
-<rh-label color="blue" icon="web-icon-alert-danger">Blue label</rh-label>
-```
-
-### Colors
-<rh-label color="blue">Blue</rh-label>
-<rh-label color="green">Green</rh-label>
-<rh-label color="orange">Orange</rh-label>
-<rh-label color="red">Red</rh-label>
-<rh-label color="purple">Purple</rh-label>
-<rh-label color="cyan">Cyan</rh-label>
-<rh-label>Grey</rh-label>
-
-{% endband %}
-
-{% renderSlots %}{% endrenderSlots %}
-
-{% renderAttributes %}{% endrenderAttributes %}
-
-{% renderMethods %}{% endrenderMethods %}
-
-{% renderEvents %}{% endrenderEvents %}
-
-{% renderCssCustomProperties %}{% endrenderCssCustomProperties %}
-
-{% renderCssParts %}{% endrenderCssParts %}
