+++ elements/rh-tag/README.md
@@ -1,11 +1,11 @@
-# Label
+# Tag
 Add a description of the component here.
 
 ## Usage
 Describe how best to use this web component along with best practices.
 
 ```html
-<rh-label>
+<rh-tag>
 
-</rh-label>
+</rh-tag>
 ```
