+++ elements/rh-tag/demo/rh-tag.js
@@ -1,0 +1,5 @@
+import '@rhds/elements/rh-tag/rh-tag.js';
+
+const root = document.querySelector('[data-demo="rh-tag"]')?.shadowRoot ?? document;
+
+root.querySelector('rh-tag');
