+++ elements/rh-tag/docs/rh-tag.md
@@ -1,0 +1,40 @@
+{% renderOverview %}
+  <rh-tag color="blue">Blue</rh-tag>
+{% endrenderOverview %}
+
+{% band header="Usage" %}
+```html
+<rh-tag color="blue">Blue</rh-tag>
+```
+{% endband %}
+
+{% band header="Variants" %}
+### With Icon
+<rh-tag color="blue" icon="web-icon-alert-danger">Blue label</rh-tag>
+
+```html
+<rh-tag color="blue" icon="web-icon-alert-danger">Blue label</rh-tag>
+```
+
+### Colors
+<rh-tag color="blue">Blue</rh-tag>
+<rh-tag color="green">Green</rh-tag>
+<rh-tag color="orange">Orange</rh-tag>
+<rh-tag color="red">Red</rh-tag>
+<rh-tag color="purple">Purple</rh-tag>
+<rh-tag color="cyan">Cyan</rh-tag>
+<rh-tag>Grey</rh-tag>
+
+{% endband %}
+
+{% renderSlots %}{% endrenderSlots %}
+
+{% renderAttributes %}{% endrenderAttributes %}
+
+{% renderMethods %}{% endrenderMethods %}
+
+{% renderEvents %}{% endrenderEvents %}
+
+{% renderCssCustomProperties %}{% endrenderCssCustomProperties %}
+
+{% renderCssParts %}{% endrenderCssParts %}
